PHP 9 - Cookies and Sessions

+++ views/film-single.tpl
@@ -9,10 +9,13 @@
         <div class="col">
             <div class="card__header">
                 <h4 class="title-4"><?=$film['name']?></h4>
-                <div class="buttons">
-                    <a href="edit.php?id=<?=$film['id']?>" class="button button--edit">Редактировать</a>
-                    <a href="index.php?action=delete&id=<?=$film['id']?>" class="button button--delete">Удалить</a>
-                </div>
+                <?php
+                    if (isAdmin()) { ?>
+                        <div class="buttons">
+                            <a href="edit.php?id=<?=$film['id']?>" class="button button--edit">Редактировать</a>
+                            <a href="index.php?action=delete&id=<?=$film['id']?>" class="button button--delete">Удалить</a>
+                        </div>
+                <?php } ?>
             </div>
             <div class="badge"><?=$film['genre']?></div>
             <div class="badge"><?=$film['year']?></div>
